+++ tests/v2/integration/setup/main.go
@@ -19,12 +19,12 @@ import (
 	"github.com/creasty/defaults"
 	provisioningv1api "github.com/rancher/rancher/pkg/apis/provisioning.cattle.io/v1"
 	v1 "github.com/rancher/rancher/pkg/apis/rke.cattle.io/v1"
-	pkgpf "github.com/rancher/rancher/pkg/portforward"
 	rancherClient "github.com/rancher/rancher/tests/framework/clients/rancher"
 	management "github.com/rancher/rancher/tests/framework/clients/rancher/generated/management/v3"
 	"github.com/rancher/rancher/tests/framework/extensions/token"
 	"github.com/rancher/rancher/tests/framework/pkg/config"
 	namegen "github.com/rancher/rancher/tests/framework/pkg/namegenerator"
+	pkgpf "github.com/rancher/rancher/tests/framework/pkg/portforward"
 	"github.com/rancher/rancher/tests/v2prov/clients"
 	"github.com/rancher/rancher/tests/v2prov/cluster"
 	testdefaults "github.com/rancher/rancher/tests/v2prov/defaults"
